0.2.10v

+++ fs/src/main/java/org/jnode/fs/fat/FatFormatter.java
@@ -77,7 +77,7 @@ public class FatFormatter {
 
     private static int calculateDefaultSectorsPerCluster(int bps, int nbTotalSectors) {
         // Apply the default cluster size from MS
-        long sizeInMB = (nbTotalSectors * bps) / (1024 * 1024);
+        long sizeInMB = ((long) nbTotalSectors * bps) / (1024 * 1024);
 
         int spc;
